Use lib-asan sysroot when ASAN is enabled
This commit also adds a unstable rustc flag to toggle using prebuilt
libraries.
Since we are not yet distributing the prebuilt artifacts, it is currently
set to `no`. Eventually, with cargo and `-Z build-std` we want to
disable using the prebuilt standard libraries (which is why it is should
later be `yes` by default).

Co-authored-by: Eric Holk <ericholk@microsoft.com>